+++ Src/Newtonsoft.Json/JsonConvert.cs
@@ -37,6 +37,7 @@ using Newtonsoft.Json.Serialization;
 using System.Text;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
+using System.Diagnostics.CodeAnalysis;
 #if HAVE_XLINQ
 using System.Xml.Linq;
 #endif
