+++ Src/Newtonsoft.Json/Utilities/DictionaryWrapper.cs
@@ -29,11 +29,11 @@ using System.Collections;
 using System.Threading;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
+using System.Diagnostics.CodeAnalysis;
 #if !HAVE_LINQ
 using Newtonsoft.Json.Utilities.LinqBridge;
 #else
 using System.Linq;
-
 #endif
 
 namespace Newtonsoft.Json.Utilities
